--- howl/src/test/kotlin/io/quartic/howl/storage/Yeah.kt
@@ -1,11 +1,0 @@
-package io.quartic.howl.storage
-
-import org.junit.Test
-import java.util.*
-
-class Yeah {
-    @Test
-    fun name() {
-        println(UUID.randomUUID())
-    }
-}
