+++ howl/src/test/kotlin/io/quartic/howl/storage/manual/ExternalIdGenerator.kt
@@ -1,0 +1,7 @@
+package io.quartic.howl.storage.manual
+
+import java.util.*
+
+fun main(args: Array<String>) {
+    println(UUID.randomUUID())
+}
