+++ howl/src/test/kotlin/io/quartic/howl/storage/manual/S3Validator.kt
@@ -1,8 +1,9 @@
-package io.quartic.howl.storage
+package io.quartic.howl.storage.manual
 
 import io.quartic.common.application.DEV_MASTER_KEY_BASE64
 import io.quartic.common.secrets.EncryptedSecret
 import io.quartic.common.secrets.SecretsCodec
+import io.quartic.howl.storage.S3StorageFactory
 import io.quartic.howl.storage.StorageCoords.Managed
 import io.quartic.howl.storage.S3StorageFactory.Config
 import org.hamcrest.Matchers.equalTo
